requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
google-generativeai>=0.8.5
django-cors-headers>=4.7.0
drf-spectacular
drf-spectacular[sidecar]